+++ test/com/lynxanalytics/biggraph/TestUtils.scala
@@ -14,6 +14,7 @@ import scala.util.Random
 object Benchmark extends Tag("Benchmark")
 object SphynxOnly extends Tag("SphynxOnly")
 object SparkOnly extends Tag("SparkOnly")
+object RequiresDocker extends Tag("RequiresDocker")
 
 object TestUtils {
   def RDDToSortedString(rdd: spark.rdd.RDD[_]): String = {